old/requirements.txt
requests
flask
python-telegram-bot
jieba
markovify
apscheduler
